pyproject.toml
[project]
name = "auto-krr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruamel-yaml>=0.19.1",
]

[project.scripts]
auto-krr = "auto_krr.cli:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
